some database changes have made

+++ ECommercial.Entities/concrete/Order.cs
@@ -5,7 +5,7 @@ namespace ECommercial.Entites.concrete
 {
     public class Order:IEntity
     {
-        public virtual short Id { get; set; }
+        public virtual long Id { get; set; }
         public virtual int UserId { get; set; }
         public virtual int ClaimAddressId { get; set; }
         public virtual int InvoiceId { get; set; }
